requirements.txt
# FastAPI & Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.4.0
pydantic-settings>=2.0.0

# HTTP Client
requests>=2.31.0
aiohttp>=3.8.0

# Environment Management
python-dotenv>=1.0.0

# AI Providers
openai>=1.3.0
anthropic>=0.7.0

# Utilities
typing-extensions>=4.8.0

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Logging and Monitoring
python-json-logger>=2.0.0

# Data Management
aiofiles>=23.0.0 
